requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.21.0
pandas>=2.0.0
plotly>=5.15.0
azure-cognitiveservices-vision-customvision>=3.1.0
google-cloud-automl>=2.11.0
google-auth>=2.20.0
matplotlib>=3.7.0
requests>=2.28.0
setuptools>=65.0.0
requests==2.31.0
